Move CachedPubSubMetadataGetter to common directory
