+++ internal/venice-common/src/main/java/com/linkedin/venice/pubsub/manager/CachedPubSubMetadataGetter.java
@@ -1,4 +1,4 @@
-package com.linkedin.davinci.kafka.consumer;
+package com.linkedin.venice.pubsub.manager;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 
@@ -6,7 +6,6 @@ import com.linkedin.venice.pubsub.PubSubTopicPartitionImpl;
 import com.linkedin.venice.pubsub.api.PubSubTopic;
 import com.linkedin.venice.pubsub.api.PubSubTopicPartition;
 import com.linkedin.venice.pubsub.api.exceptions.PubSubTopicDoesNotExistException;
-import com.linkedin.venice.pubsub.manager.TopicManager;
 import com.linkedin.venice.stats.StatsErrorCode;
 import com.linkedin.venice.utils.concurrent.VeniceConcurrentHashMap;
 import java.util.Map;
@@ -22,7 +21,7 @@ import org.apache.logging.log4j.Logger;
  * Because get real-time topic offset, get producer timestamp, and check topic existence are expensive, so we will only
  * retrieve such information after the predefined ttlMs
  */
-class CachedPubSubMetadataGetter {
+public class CachedPubSubMetadataGetter {
   private static final Logger LOGGER = LogManager.getLogger(CachedPubSubMetadataGetter.class);
   private static final int DEFAULT_MAX_RETRY = 10;
 
@@ -31,7 +30,7 @@ class CachedPubSubMetadataGetter {
   private final Map<PubSubMetadataCacheKey, ValueAndExpiryTime<Long>> offsetCache;
   private final Map<PubSubMetadataCacheKey, ValueAndExpiryTime<Long>> lastProducerTimestampCache;
 
-  CachedPubSubMetadataGetter(long timeToLiveMs) {
+  public CachedPubSubMetadataGetter(long timeToLiveMs) {
     this.ttlNs = MILLISECONDS.toNanos(timeToLiveMs);
     this.topicExistenceCache = new VeniceConcurrentHashMap<>();
     this.offsetCache = new VeniceConcurrentHashMap<>();
@@ -43,7 +42,7 @@ class CachedPubSubMetadataGetter {
    * return the next available offset rather the latest used offset. Therefore,
    * the value will be 1 offset greater than what's expected.
    */
-  long getOffset(TopicManager topicManager, PubSubTopic pubSubTopic, int partitionId) {
+  public long getOffset(TopicManager topicManager, PubSubTopic pubSubTopic, int partitionId) {
     final String sourcePubSubServer = topicManager.getPubSubClusterAddress();
     PubSubTopicPartition pubSubTopicPartition = new PubSubTopicPartitionImpl(pubSubTopic, partitionId);
     try {
@@ -59,7 +58,7 @@ class CachedPubSubMetadataGetter {
     }
   }
 
-  long getEarliestOffset(TopicManager topicManager, PubSubTopicPartition pubSubTopicPartition) {
+  public long getEarliestOffset(TopicManager topicManager, PubSubTopicPartition pubSubTopicPartition) {
     final String sourcePubSubServer = topicManager.getPubSubClusterAddress();
     try {
       return fetchMetadata(
@@ -74,7 +73,9 @@ class CachedPubSubMetadataGetter {
     }
   }
 
-  long getProducerTimestampOfLastDataMessage(TopicManager topicManager, PubSubTopicPartition pubSubTopicPartition) {
+  public long getProducerTimestampOfLastDataMessage(
+      TopicManager topicManager,
+      PubSubTopicPartition pubSubTopicPartition) {
     try {
       return fetchMetadata(
           new PubSubMetadataCacheKey(topicManager.getPubSubClusterAddress(), pubSubTopicPartition),
@@ -87,7 +88,7 @@ class CachedPubSubMetadataGetter {
     }
   }
 
-  boolean containsTopic(TopicManager topicManager, PubSubTopic pubSubTopic) {
+  public boolean containsTopic(TopicManager topicManager, PubSubTopic pubSubTopic) {
     return fetchMetadata(
         new PubSubMetadataCacheKey(
             topicManager.getPubSubClusterAddress(),
